Use _geeglm naming
This was a long-ago change and I'm not exactly sure what was the purpose. But it seems like this is to use the _geeglm naming for some functions, such as vcov -> vcov_geeglm, and model.matrix -> model.matrix_geeglm. Note sure if this is related to the geepack update a few months ago.